chore(AttributesModal): move to containers [YTFRONT-5665]
