+++ packages/ui/src/ui/containers/AttributesModal/AttributesModal.scss
